ready application add status and async shelljs

+++ collections/bundles.js
@@ -11,25 +11,41 @@ Bundles = new FS.Collection("bundles", {
 });
 
 isServer(() => {
-  Bundles.on('stored', (file, storeName) => {
+  Bundles.on('stored', Meteor.bindEnvironment((file, storeName) => {
     const tar = `${file.collectionName}-${file._id}-${file.name()}`;
 
     // CD BUNDLE DIR
     shell.cd(BUNDLE_DIR);
 
-    // GENERATE BUNDLE DIR
-    shell.mkdir(file._id);
+    // IF NOT EXISTS DIR
+    if (!shell.test('-e', file._id)) {
+      // GENERATE BUNDLE DIR
+      shell.mkdir(file._id);
 
-    // EXTRACT APPLICATION
-    shell.exec(`tar -xf ${tar} -C ${file._id} --strip 1`, { silent: true });
+      // EXTRACT APPLICATION
+      shell.exec(`tar -xf ${tar} -C ${file._id} --strip 1`, { silent: true });
 
-    // REMOVE
-    shell.rm('-rf', tar);
+      // REMOVE
+      shell.rm('-rf', tar);
 
-    // CD SERVER PACKAGES
-    shell.cd(`${file._id}/programs/server`);
+      // CD SERVER PACKAGES
+      shell.cd(`${file._id}/programs/server`);
 
-    // NPM PACKAGES INSTALL
-    shell.exec('npm install', { silent: true });
-  });
+      // NPM PACKAGES INSTALL
+      const child = shell.exec('npm install', { silent: true, async: true  });
+
+      /* || bindEnvironment
+       *
+       * if npm install completed then update
+       * application status READY..
+       */
+      child.stdout.on('end', Meteor.bindEnvironment(() => {
+        Applications.update({ bundleId: file._id }, {
+          $set: {
+            status: STATUS_ALLOWED_VALUES[3] // READY
+          }
+        });
+      }));
+    }
+  }));
 });
